Added Protobuf into MessagingService

+++ services/messageservice/src/main/java/com/shop/messageservice/dto/OrderItemDTO.java
@@ -1,5 +1,7 @@
 package com.shop.messageservice.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class OrderItemDTO {
     private String name;
     private int quantity;
-    private double pricePerUnit;
+    private BigDecimal pricePerUnit;
 }
