+++ services/paymentservice/src/main/java/com/shop/paymentservice/strategy/PaymentStrategy.java
@@ -1,11 +1,13 @@
 package com.shop.paymentservice.strategy;
 
+import java.math.BigDecimal;
+
 import com.shop.paymentservice.dto.PaymentRequest;
 import com.shop.paymentservice.dto.PaymentResponse;
 
 public interface PaymentStrategy {
 	PaymentResponse processPayment(PaymentRequest request);
-	default boolean processRefund(String transactionId, double refundAmount) {
+	default boolean processRefund(String transactionId, BigDecimal refundAmount) {
 		return false;
 	}
 }
